Additional keys on thumb cluster and rationalize RGB
Add a SHIFT on inner left thumb, which really helps with ALL CAPS, and
add TAB and GRAVE to the big red arrows, which gives an altenative
route for app and window switching. Also, remove the bindings for the
bottom row mods that I have physically removed the keys of. And make
the home row rgb colors be consistent with their respective modifiers.